+++ accounts/keystore/keystore.go
@@ -34,8 +34,8 @@ import (
 	"github.com/theQRL/go-zond/accounts"
 	"github.com/theQRL/go-zond/common"
 	"github.com/theQRL/go-zond/core/types"
+	"github.com/theQRL/go-zond/crypto/pqcrypto"
 	"github.com/theQRL/go-zond/event"
-	"github.com/theQRL/go-zond/pqcrypto"
 )
 
 var (
