+++ cmd/evm/internal/t8ntool/transition.go
@@ -28,14 +28,14 @@ import (
 	"github.com/theQRL/go-qrllib/dilithium"
 	"github.com/theQRL/go-zond/common"
 	"github.com/theQRL/go-zond/common/hexutil"
-	"github.com/theQRL/go-zond/consensus/misc"
+	"github.com/theQRL/go-zond/consensus/misc/eip1559"
 	"github.com/theQRL/go-zond/core"
 	"github.com/theQRL/go-zond/core/state"
 	"github.com/theQRL/go-zond/core/types"
 	"github.com/theQRL/go-zond/core/vm"
+	"github.com/theQRL/go-zond/crypto/pqcrypto"
 	"github.com/theQRL/go-zond/log"
 	"github.com/theQRL/go-zond/params"
-	"github.com/theQRL/go-zond/pqcrypto"
 	"github.com/theQRL/go-zond/rlp"
 	"github.com/theQRL/go-zond/tests"
 	"github.com/theQRL/go-zond/zond/tracers/logger"
@@ -276,8 +276,8 @@ func signUnsignedTransactions(txs []*txWithKey, signer types.Signer) (types.Tran
 	for i, tx := range txs {
 		var (
 			signature = tx.tx.RawSignatureValues()
-			signed  *types.Transaction
-			err     error
+			signed    *types.Transaction
+			err       error
 		)
 		if tx.key == nil || signature != nil {
 			// Already signed
